+++ src/app/provider/index.tsx
@@ -6,8 +6,8 @@ import {
   QueryClientProvider,
 } from '@tanstack/react-query';
 import theme from '../../utils/theme/theme';
-import { CurrentUserContext, CurrentUserContextObjectType } from '@/contexts/current-user.context';
-import React, { useEffect } from 'react';
+import React from 'react';
+import { CurrentUserProvider } from '@/contexts/current-user.context';
 
 export default function Provider({
   children,
@@ -16,16 +16,12 @@ export default function Provider({
 }) {
   const queryClient = new QueryClient();
 
-   // put in context provider location
-  const [currentUser, setCurrentUser] = React.useState<CurrentUserContextObjectType | null>(null);
-
-
   return (
     <QueryClientProvider client={queryClient}>
       <ChakraProvider theme={theme}>
-        <CurrentUserContext.Provider value={{currentUser, setCurrentUser}}>
+        <CurrentUserProvider>
           {children}
-        </CurrentUserContext.Provider>
+        </CurrentUserProvider>
       </ChakraProvider>
     </QueryClientProvider>
   );
