+++ src/components/ProtectedRoute.tsx
@@ -2,19 +2,30 @@
 
 
 import { useProtectedRoute } from "@/hooks/routes.hooks";
-import React from "react";
+import { useRouter } from "next/navigation";
+import React, { useEffect } from "react";
 
 
 export default function ProtectedRoute({ children }: {
     children: React.ReactNode;
   }) {
+  const router = useRouter();
   const isValidRoute = useProtectedRoute();
 
-  if (!isValidRoute) {
-    return <div>Loading...</div>;
+  useEffect(() => {
+    if (isValidRoute === false){
+      router.replace("/");
+    }
+  }, []);
+
+
+  if (isValidRoute){
+    return (
+      <>{children}</>
+    );
   }
 
-  return (
-    <>{children}</>
-  );
+  return <div>Loading...</div>;
+
+  
 }
